rename lib -> Library
