+++ Library/PatchUrlResolver.rb
